Fix rsyncOpts variable and add delete option
Valid values for delete are:

delete	=	true	Default. Lsyncd will delete on the target whatever is not in the source. At startup and whats being deleted during normal operation.
delete	=	false	Lsyncd will not delete any files on the target. Not on and startup nor on normal operation. (Overwrites are possible though)
delete	=	'startup'	Lsyncd will delete files on the target when it starts up but not on normal operation.
delete	=	'running'	Lsyncd will not delete files on the target when it starts up but will delete those that are removed during normal operation.